requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
idna==3.11
lxml==6.0.2
numpy==2.3.4
pandas==2.3.3
pillow==12.0.0
python-dateutil==2.9.0.post0
pytrends==4.9.2
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
